requirements.txt
Flask
gspread
oauth2client
gunicorn
python-dotenv
python-amazon-paapi @ git+https://github.com/sergioteula/python-amazon-paapi.git@master
